feat: 23 - refactor registration
